+++ sdk/appcenter-crashes/src/androidTest/java/com/microsoft/appcenter/crashes/WrapperSdkExceptionManagerAndroidTest.java
@@ -49,7 +49,9 @@ public class WrapperSdkExceptionManagerAndroidTest {
     public void setUp() {
         android.util.Log.i(TAG, "Cleanup");
         SharedPreferencesManager.clear();
-        for (File logFile : ErrorLogHelper.getErrorStorageDirectory().listFiles()) {
+        File[] files = ErrorLogHelper.getErrorStorageDirectory().listFiles();
+        assertNotNull(files);
+        for (File logFile : files) {
             if (!logFile.isDirectory()) {
                 assertTrue(logFile.delete());
             }
@@ -72,6 +74,18 @@ public class WrapperSdkExceptionManagerAndroidTest {
         method.setAccessible(true);
         method.invoke(appCenter, mock(Channel.class));
 
+        /*
+         * Since this is a real Android test, it might actually try to send crash logs
+         * and will delete files on sending completion. Avoid that by requesting user confirmation.
+         */
+        Crashes.setListener(new AbstractCrashesListener() {
+
+            @Override
+            public boolean shouldAwaitUserConfirmation() {
+                return false;
+            }
+        });
+
         /* Start crashes. */
         AppCenter.start(Crashes.class);
 
@@ -83,20 +97,23 @@ public class WrapperSdkExceptionManagerAndroidTest {
     public void saveWrapperException() throws java.lang.Exception {
 
         class ErrorData {
-            private byte[] data;
+            private String data;
             private UUID id;
         }
 
         ErrorData errorA = new ErrorData();
-        errorA.data = new byte[]{};
+        errorA.data = "";
 
         ErrorData errorB = new ErrorData();
         errorB.data = null;
 
         ErrorData errorC = new ErrorData();
-        errorC.data = new byte[]{'d', 'a', 't', 'a'};
+        errorC.data = "data";
 
-        ErrorData[] errors = new ErrorData[]{errorA, errorB, errorC};
+        ErrorData errorD = new ErrorData();
+        errorD.data = "otherData";
+
+        ErrorData[] errors = new ErrorData[]{errorA, errorB, errorC, errorD};
 
         for (ErrorData error : errors) {
 
@@ -105,7 +122,7 @@ public class WrapperSdkExceptionManagerAndroidTest {
 
             /* Save crash. */
             error.id = WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), null, new Exception(), error.data);
-            byte[] loadedData = WrapperSdkExceptionManager.loadWrapperExceptionData(error.id);
+            String loadedData = WrapperSdkExceptionManager.loadWrapperExceptionData(error.id);
 
             if (error.data == null) {
                 assertNull(loadedData);
@@ -113,28 +130,26 @@ public class WrapperSdkExceptionManagerAndroidTest {
             }
 
             assertNotNull(loadedData);
-            for (int i = 0; i < error.data.length; ++i) {
-                assertEquals(error.data[i], loadedData[i]);
-            }
+            assertEquals(error.data, loadedData);
         }
 
-        /* Even after deleting errorA, it should exist in memory - so, we can still load it. */
+        /* Even after deleting errorA and errorD, they should exist in memory - so, we can still load it. */
         WrapperSdkExceptionManager.deleteWrapperExceptionData(errorA.id);
-        byte[] loadedDataA = WrapperSdkExceptionManager.loadWrapperExceptionData(errorA.id);
+        String loadedDataA = WrapperSdkExceptionManager.loadWrapperExceptionData(errorA.id);
         assertNotNull(loadedDataA);
-        for (int i = 0; i < errorA.data.length; ++i) {
-            assertEquals(errorA.data[i], loadedDataA[i]);
-        }
+        assertEquals(errorA.data, loadedDataA);
+        WrapperSdkExceptionManager.deleteWrapperExceptionData(errorD.id);
+        String loadedDataD = WrapperSdkExceptionManager.loadWrapperExceptionData(errorD.id);
+        assertNotNull(loadedDataD);
+        assertEquals(errorD.data, loadedDataD);
 
         /* Try to load data bypassing the cache. */
         WrapperSdkExceptionManager.sWrapperExceptionDataContainer.clear();
-        byte[] loadedDataC = WrapperSdkExceptionManager.loadWrapperExceptionData(errorC.id);
+        String loadedDataC = WrapperSdkExceptionManager.loadWrapperExceptionData(errorC.id);
         assertNotNull(loadedDataC);
-        for (int i = 0; i < errorC.data.length; ++i) {
-            assertEquals(errorC.data[i], loadedDataC[i]);
-        }
+        assertEquals(errorC.data, loadedDataC);
 
         /* Save another crash without reset: will be ignored as only 1 crash per process. */
-        assertNull(WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), null, new Exception(), new byte[]{'e'}));
+        assertNull(WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), null, new Exception(), "e"));
     }
 }