+++ sdk/appcenter-crashes/src/test/java/com/microsoft/appcenter/crashes/WrapperSdkExceptionManagerTest.java
@@ -38,6 +38,7 @@ import java.io.IOException;
 import java.util.Map;
 import java.util.UUID;
 
+import static android.util.Log.getStackTraceString;
 import static com.microsoft.appcenter.utils.PrefStorageConstants.KEY_ENABLED;
 import static org.junit.Assert.assertNull;
 import static org.mockito.Matchers.any;
@@ -62,6 +63,8 @@ import static org.powermock.api.mockito.PowerMockito.whenNew;
 public class WrapperSdkExceptionManagerTest {
 
     private static final String CRASHES_ENABLED_KEY = KEY_ENABLED + "_" + Crashes.getInstance().getServiceName();
+
+    private static final String STACK_TRACE = "Sample stack trace";
 
     @Rule
     public final PowerMockRule rule = new PowerMockRule();
@@ -116,11 +119,6 @@ public class WrapperSdkExceptionManagerTest {
         File file = mock(File.class);
         whenNew(File.class).withAnyArguments().thenReturn(file);
         when(file.exists()).thenReturn(true);
-        doThrow(new IOException()).when(FileManager.class);
-        FileManager.readObject(any(File.class));
-        assertNull(WrapperSdkExceptionManager.loadWrapperExceptionData(UUID.randomUUID()));
-        doThrow(new ClassNotFoundException()).when(FileManager.class);
-        FileManager.readObject(any(File.class));
         assertNull(WrapperSdkExceptionManager.loadWrapperExceptionData(UUID.randomUUID()));
         assertNull(WrapperSdkExceptionManager.loadWrapperExceptionData(null));
     }
@@ -166,56 +164,62 @@ public class WrapperSdkExceptionManagerTest {
         LogSerializer logSerializer = Mockito.mock(LogSerializer.class);
         when(logSerializer.serializeLog(any(ManagedErrorLog.class))).thenReturn("mock");
         Crashes.getInstance().setLogSerializer(logSerializer);
-        byte[] data = new byte[]{'d'};
+        String data = "d";
         WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), null, new Exception(), data);
         verifyStatic();
-        FileManager.writeObject(any(File.class), eq(data));
+        FileManager.write(any(File.class), eq(data));
 
         /* We can't do it twice in the same process. */
-        data = new byte[]{'e'};
+        data = "e";
         WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), null, new Exception(), data);
         verifyStatic(never());
-        FileManager.writeObject(any(File.class), eq(data));
+        FileManager.write(any(File.class), eq(data));
     }
 
     @Test
+    @PrepareForTest(android.util.Log.class)
     public void saveWrapperSdkCrashWithJavaThrowable() throws JSONException, IOException {
         LogSerializer logSerializer = Mockito.mock(LogSerializer.class);
         when(logSerializer.serializeLog(any(ManagedErrorLog.class))).thenReturn("mock");
         Crashes.getInstance().setLogSerializer(logSerializer);
-        byte[] data = new byte[]{'d'};
+        String data = "d";
         Throwable throwable = new Throwable();
+        mockStatic(android.util.Log.class);
+        Mockito.when(getStackTraceString(any(Throwable.class))).thenReturn(STACK_TRACE);
         WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), throwable, new Exception(), data);
         verifyStatic();
-        FileManager.writeObject(any(File.class), eq(data));
+        FileManager.write(any(File.class), eq(data));
         verifyStatic();
-        FileManager.writeObject(any(File.class), eq(throwable));
+        FileManager.write(any(File.class), eq(STACK_TRACE));
 
         /* We can't do it twice in the same process. */
-        data = new byte[]{'e'};
+        data = "e";
         WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), throwable, new Exception(), data);
         verifyStatic(never());
-        FileManager.writeObject(any(File.class), eq(data));
+        FileManager.write(any(File.class), eq(data));
         verifyStatic();
-        FileManager.writeObject(any(File.class), eq(throwable));
+        FileManager.write(any(File.class), eq(STACK_TRACE));
     }
 
     @Test
+    @PrepareForTest({android.util.Log.class})
     public void saveWrapperSdkCrashWithOnlyJavaThrowable() throws JSONException, IOException {
         LogSerializer logSerializer = Mockito.mock(LogSerializer.class);
         when(logSerializer.serializeLog(any(ManagedErrorLog.class))).thenReturn("mock");
         Crashes.getInstance().setLogSerializer(logSerializer);
         Throwable throwable = new Throwable();
+        mockStatic(android.util.Log.class);
+        Mockito.when(getStackTraceString(any(Throwable.class))).thenReturn(STACK_TRACE);
         WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), throwable, new Exception(), null);
         verifyStatic(never());
-        FileManager.writeObject(any(File.class), isNull(byte[].class));
+        FileManager.write(any(File.class), isNull(String.class));
         verifyStatic();
-        FileManager.writeObject(any(File.class), eq(throwable));
+        FileManager.write(any(File.class), eq(STACK_TRACE));
 
         /* We can't do it twice in the same process. */
         WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), throwable, new Exception(), null);
         verifyStatic();
-        FileManager.writeObject(any(File.class), eq(throwable));
+        FileManager.write(any(File.class), eq(STACK_TRACE));
     }
 
     @Test
@@ -232,7 +236,7 @@ public class WrapperSdkExceptionManagerTest {
             }
         })).thenReturn(throwableFile);
         when(throwableFile.createNewFile()).thenReturn(false);
-        byte[] data = new byte[]{'d'};
+        String data = "d";
         WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), null, new Exception(), data);
         verifyStatic();
         AppCenterLog.error(anyString(), anyString(), argThat(new ArgumentMatcher<Throwable>() {
@@ -244,7 +248,7 @@ public class WrapperSdkExceptionManagerTest {
         }));
 
         /* Second call is ignored. */
-        data = new byte[]{'e'};
+        data = "e";
         WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), null, new Exception(), data);
 
         /* No more error. */
@@ -263,7 +267,8 @@ public class WrapperSdkExceptionManagerTest {
         LogSerializer logSerializer = Mockito.mock(LogSerializer.class);
         when(logSerializer.serializeLog(any(ManagedErrorLog.class))).thenThrow(new JSONException("mock"));
         Crashes.getInstance().setLogSerializer(logSerializer);
-        WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), null, new Exception(), new byte[]{'d'});
+        String data = "d";
+        WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), null, new Exception(), data);
         verifyStatic();
         AppCenterLog.error(anyString(), anyString(), argThat(new ArgumentMatcher<Throwable>() {
 
@@ -274,7 +279,8 @@ public class WrapperSdkExceptionManagerTest {
         }));
 
         /* Second call is ignored. */
-        WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), null, new Exception(), new byte[]{'e'});
+        data = "e";
+        WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), null, new Exception(), data);
 
         /* No more error. */
         verifyStatic();
@@ -294,7 +300,8 @@ public class WrapperSdkExceptionManagerTest {
         LogSerializer logSerializer = Mockito.mock(LogSerializer.class);
         when(logSerializer.serializeLog(any(ManagedErrorLog.class))).thenReturn("mock");
         Crashes.getInstance().setLogSerializer(logSerializer);
-        WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), null, new Exception(), new byte[]{'d'});
+        String data = "d";
+        WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), null, new Exception(), data);
         verifyStatic();
         AppCenterLog.error(anyString(), anyString(), argThat(new ArgumentMatcher<Throwable>() {
 
@@ -305,7 +312,8 @@ public class WrapperSdkExceptionManagerTest {
         }));
 
         /* Second call is ignored. */
-        WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), null, new Exception(), new byte[]{'e'});
+        data = "e";
+        WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), null, new Exception(), data);
 
         /* No more error. */
         verifyStatic();
@@ -320,9 +328,9 @@ public class WrapperSdkExceptionManagerTest {
 
     @Test
     public void saveWrapperSdkCrashFailsWithIOExceptionAfterLog() throws IOException, JSONException {
-        byte[] data = {'d'};
+        String data = "d";
         doThrow(new IOException()).when(FileManager.class);
-        FileManager.writeObject(any(File.class), eq(data));
+        FileManager.write(any(File.class), eq(data));
         LogSerializer logSerializer = Mockito.mock(LogSerializer.class);
         when(logSerializer.serializeLog(any(ManagedErrorLog.class))).thenReturn("mock");
         Crashes.getInstance().setLogSerializer(logSerializer);
@@ -337,7 +345,8 @@ public class WrapperSdkExceptionManagerTest {
         }));
 
         /* Second call is ignored. */
-        WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), null, new Exception(), new byte[]{'e'});
+        data = "e";
+        WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), null, new Exception(), data);
 
         /* No more error. */
         verifyStatic();
@@ -355,7 +364,8 @@ public class WrapperSdkExceptionManagerTest {
         when(SharedPreferencesManager.getBoolean(CRASHES_ENABLED_KEY, true)).thenReturn(false);
         LogSerializer logSerializer = Mockito.mock(LogSerializer.class);
         Crashes.getInstance().setLogSerializer(logSerializer);
-        WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), null, new Exception(), new byte[]{'d'});
+        String data = "d";
+        WrapperSdkExceptionManager.saveWrapperException(Thread.currentThread(), null, new Exception(), data);
         verify(logSerializer, never()).serializeLog(any(Log.class));
         verifyNoMoreInteractions(ErrorLogHelper.class);
     }